+++ clinicaOdontologica/src/main/java/com/backend/clinicaOdontologica/dto/entrada/TurnoEntradaDto.java
@@ -5,10 +5,10 @@ import java.time.LocalDateTime;
 
 public class TurnoEntradaDto {
     @NotNull(message = "El ID del odontólogo no puede ser nulo")
-    private int idOdontologo;
+    private Long idOdontologo;
 
     @NotNull(message = "El ID del paciente no puede ser nulo")
-    private int idPaciente;
+    private Long idPaciente;
 
     @NotNull(message = "La fecha y hora del turno no puede ser nula")
     private LocalDateTime fechaYHora;
@@ -16,25 +16,25 @@ public class TurnoEntradaDto {
     public TurnoEntradaDto() {
     }
 
-    public TurnoEntradaDto(int idOdontologo, int idPaciente, LocalDateTime fechaYHora) {
+    public TurnoEntradaDto(Long idOdontologo, Long idPaciente, LocalDateTime fechaYHora) {
         this.idOdontologo = idOdontologo;
         this.idPaciente = idPaciente;
         this.fechaYHora = fechaYHora;
     }
 
-    public int getIdOdontologo() {
+    public Long getIdOdontologo() {
         return idOdontologo;
     }
 
-    public void setIdOdontologo(int idOdontologo) {
+    public void setIdOdontologo(Long idOdontologo) {
         this.idOdontologo = idOdontologo;
     }
 
-    public int getIdPaciente() {
+    public Long getIdPaciente() {
         return idPaciente;
     }
 
-    public void setIdPaciente(int idPaciente) {
+    public void setIdPaciente(Long idPaciente) {
         this.idPaciente = idPaciente;
     }
 
@@ -45,4 +45,5 @@ public class TurnoEntradaDto {
     public void setFechaYHora(LocalDateTime fechaYHora) {
         this.fechaYHora = fechaYHora;
     }
+
 }
